requirements.txt
Django==4.2.0
psycopg==3.1.9
pillow==9.5.0
djangorestframework==3.14.0
